etl/requirements.txt
osmium==3.7.0
pandas==2.3.2
psycopg2-binary==2.9.10
pyproj==3.7.2
python-dotenv==1.1.1
requests==2.32.5
shapely==2.1.1
tqdm==4.67.1
